+++ spring-cloud-tencent-examples/quickstart-example/quickstart-caller-service/src/main/java/com/tencent/cloud/quickstart/caller/CustomMetadataProvider.java
@@ -17,18 +17,18 @@
 
 package com.tencent.cloud.quickstart.caller;
 
+import com.tencent.cloud.common.spi.InstanceMetadataProvider;
+
 import org.springframework.stereotype.Component;
 
 /**
- * Quickstart callee feign client fallback.
- *
- * @author Haotian Zhang
+ *@author lepdou 2022-06-29
  */
 @Component
-public class QuickstartCalleeServiceFallback implements QuickstartCalleeService {
+public class CustomMetadataProvider implements InstanceMetadataProvider {
 
 	@Override
-	public int sum(int value1, int value2) {
-		return 0;
+	public String getRegion() {
+		return "huadong";
 	}
 }